+++ SplunkAppForWazuh/appserver/static/js/directives/wz-menu/wz-menu.js
@@ -9,7 +9,8 @@
  *
  * Find more information about this on the LICENSE file.
  */
-define(['../module'], function(directives) {
+define(['../module','splunkjs/mvc/simpleform/input/dropdown', '../../services/visualizations/inputs/dropdown-input', 'splunkjs/mvc'], 
+function(directives, Dropdown, DropdownViz, mvc) {
   'use strict'
   directives.directive('wzMenu', function(BASE_URL) {
     return {
@@ -17,7 +18,10 @@ define(['../module'], function(directives) {
         $scope,
         $currentDataService,
         $navigationService,
-        $state
+        $state,
+        $notificationService,
+        $urlTokenModel,
+        $window
       ) {
         $scope.logoUrl =
           BASE_URL +
@@ -45,6 +49,139 @@ define(['../module'], function(directives) {
           $scope.openDiscover(data)
         })
 
+        let dropdownAPI;
+        let dropdownIndex;
+        let dropdownSourceType;
+
+        let onChangeListeners = [];
+        
+        const onChangeDropdownAPI = () => {
+          onChangeListeners.push(dropdownAPI.on('change', newValue => {
+            try {
+              if (newValue && $scope.currentAPI._key != newValue) {
+                selectAPI(newValue)
+              }
+            } catch (error) {
+              $notificationService.showErrorToast(error)
+            }
+          }))
+        }
+
+  
+        const onChangeDropdownIndex = () => {          
+          const dropdownInstance = dropdownIndex.getElement()
+          onChangeListeners.push(dropdownInstance.on('change', newValue => {
+            try {
+              if (newValue && dropdownInstance && $scope.currentIndex != newValue) {
+                $currentDataService.setIndex(newValue)
+                $urlTokenModel.handleValueChange(dropdownInstance)
+                $scope.currentIndex = newValue;
+                $window.location.reload();
+              }
+            } catch (error) {
+              notificationService.showErrorToast(error)
+            }
+          }))
+        }
+
+      const onChangeDropdownSourceType = () => {
+          const dropdownInstance = dropdownSourceType.getElement()
+          onChangeListeners.push(dropdownInstance.on('change', newValue => {
+              try {
+                  if (newValue && dropdownInstance && $scope.currentSourceType != newValue) {
+                    $currentDataService.setSourceType(newValue)
+                    $scope.currentSourceType = newValue       
+                    $urlTokenModel.handleValueChange(dropdownInstance)
+                    $window.location.reload();
+                  }
+              } catch (error) {
+                  $notificationService.showErrorToast(error)
+              }
+          }))
+      }
+
+        const renderDropdownAPI = () => {
+          if (dropdownAPI){
+            mvc.Components.revokeInstance(dropdownAPI.id);
+            $(`#menuSelectAPI`).html('');
+          }
+
+          dropdownAPI = new Dropdown(
+            {
+              id: `selectAPI`,
+              choices: $scope.apiList.map((item)=> ({ label:item.managerName, value:item._key })),
+              value: $scope.currentAPI._key,
+              selectFirstChoice: false,                    
+              el: $(`#menuSelectAPI`)
+            },
+            { tokens: false}
+          ).render()
+        }
+
+        const renderDropdownIndex = () => {
+          if (dropdownIndex){
+            dropdownIndex.destroy();
+            $(`#menuSelectIndex`).html('');
+          }
+
+          dropdownIndex = new DropdownViz(
+            'menuSelectIndex',
+            `| metasearch index=* sourcetype=*wazuh* | stats count by index, sourcetype | fields index`,
+            'index',
+            '$form.index$',
+            'menuSelectIndex',
+            $scope,
+            $scope.currentIndex,
+            '2017-03-14T10:0:0',
+            'now'
+          )
+        }
+
+        const renderDropdownSourceType = () => {
+          if (dropdownSourceType){
+            dropdownSourceType.destroy();
+            $(`#menuSelectSourceType`).html('');
+          }
+          dropdownSourceType = new DropdownViz(
+            'menuSelectSourceType',
+            `| metasearch index=${$scope.currentIndex} sourcetype=* | stats count by index, sourcetype | fields sourcetype`,
+            'sourcetype',
+            '$form.sourcetype$',
+            'menuSelectSourceType',
+            $scope,
+            $scope.currentSourceType,
+            '2017-03-14T10:0:0',
+            'now'
+          )
+        }
+
+        const clearListeners = () => {
+          onChangeListeners.forEach(instance => instance.stopListening())
+          onChangeListeners = []
+        }
+
+        const init = async() => {
+          update();
+          $scope.$on('$destroy', () => {
+            clearListeners()
+            dropdownAPI.destroy();
+            dropdownIndex.destroy();
+            dropdownSourceType.destroy();            
+          })
+        }
+
+        const selectAPI = async (key) => {
+          try {
+            // checking if the api is up
+            await $currentDataService.checkApiConnection(key);
+            // Selecting API
+            await $currentDataService.chose(key);
+            $window.location.reload();
+          } catch (err) {
+            $notificationService.showErrorToast(err || 'Could not select API')
+          }
+        }
+
         const checkLastState = (prefix, state) => {
           try {
             const lastState = $navigationService.getLastState()
@@ -64,13 +201,17 @@ define(['../module'], function(directives) {
           }
         }
 
-        const update = () => {
+        const update = async() => {
           try {
+            clearListeners()
+            $scope.apiList = await $currentDataService.getApiList();
             const index = $currentDataService.getIndex()
+            const sourceType = $currentDataService.getSourceType()
             const api = $currentDataService.getApi()
             $scope.currentIndex = !index ? 'wazuh' : index.index
-            $scope.currentAPI = !api ? '---' : api.managerName
-            $scope.theresAPI = $scope.currentAPI === '---' ? false : true
+            $scope.currentSourceType = !sourceType ? '*' : sourceType.sourceType
+            $scope.currentAPI = !api ? {managerName:'---', _key:'-'} : api
+            $scope.theresAPI = !!api
 
             if (checkLastState('ow-', 'overview')) {
               $scope.menuNavItem = 'overview'
@@ -82,9 +223,18 @@ define(['../module'], function(directives) {
               $scope.menuNavItem = 'settings'
             } else if (checkLastState('dev-tools', 'dev-tools')) {
               $scope.menuNavItem = 'dev-tools'
-            } else if (checkLastState('discover', 'discover')) {
+            } else if (checkLastState('discover', 'discover')) { 
               $scope.menuNavItem = 'discover'
             }
+
+            if ($scope.theresAPI && $scope.apiList.length > 1) {
+              renderDropdownAPI();
+              onChangeDropdownAPI();
+            }
+            renderDropdownIndex();
+            renderDropdownSourceType();
+            onChangeDropdownIndex();
+            onChangeDropdownSourceType();
             $scope.$applyAsync()
           } catch (error) {
             $state.go('settings.api')
@@ -102,6 +252,8 @@ define(['../module'], function(directives) {
           $scope.select(data)
           update()
         })
+
+        init()
       },
       templateUrl:
         BASE_URL +